devops/11-plataforma-interna/requirements.txt
# Dependências do Módulo 11 — Plataforma Interna
# Instalar com: pip install -r requirements.txt

# Para scripts didáticos
pyyaml==6.0.2
rich==13.9.3
requests==2.32.3
jinja2==3.1.4
pydantic==2.9.2
python-dateutil==2.9.0.post0

# Opcional: interface CLI/dashboard
click==8.1.7

# Testes
pytest==8.3.3
